requirements.txt
fdk>=0.1.60
oci
oracledb
PyYAML
pandas
pyarrow
